fix(GC)build error bump gt dependencie

+++ src/main/java/micdoodle8/mods/galacticraft/core/items/ItemOxygenTank.java
@@ -21,7 +21,7 @@ public class ItemOxygenTank extends Item
         this.setMaxStackSize(1);
 	double factor = 2.0;
 	// Config modifier goes here if anyone wants it.
-        this.setMaxDamage(Math.round(Math.pow(factor, tier - 1) * 10) * 100);
+        this.setMaxDamage((int) (Math.round(Math.pow(factor, tier - 1) * 10) * 100));
         this.setUnlocalizedName(assetName);
         this.setTextureName(GalacticraftCore.TEXTURE_PREFIX + assetName);
         this.setNoRepair();
